feat: add Trade apis
